+++ CHANGELOG.md
@@ -17,8 +17,10 @@
 - Changed increased the durability of Crumble Horn, Peacock Fan, Ore Magnet, and Giant Pick from TForest like GT6.
 - Added some processing recipes for TForest in GTC and IC2C machines.
 - Added ability for GT Scanners to be nuclear thermometers as well as EU scanners
+- Added Steam Reactor internal tanks info to GT Scanner.
 - Changed Beryllium from a fluid to a solid as it should be, recipe to convert from fluid to solid in the Centrifuge.
 - Changed the Dragon Egg Energy Siphon to a Trophy Energy Siphon which can make power from more sources (configurable).
+- Added a special new crop!
 
 # 1.1.5
 - Added ability to add materials to GTC via Craft Tweaker thanks to Trinsdar.